pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "greenlet>=3.3.2",
    "nanoid>=2.0.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
    "validator>=0.7.1",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
